Refactor `Footer` component
